Input validation

+++ src/main/java/cz/svjis/servlet/cmd/BuildingUnitDeleteCmd.java
@@ -9,6 +9,7 @@ import cz.svjis.bean.BuildingDAO;
 import cz.svjis.bean.BuildingUnit;
 import cz.svjis.servlet.CmdContext;
 import cz.svjis.servlet.Command;
+import cz.svjis.validator.Validator;
 import javax.servlet.RequestDispatcher;
 
 /**
@@ -24,14 +25,32 @@ public class BuildingUnitDeleteCmd extends Command {
     @Override
     public void execute() throws Exception {
         BuildingDAO buildingDao = new BuildingDAO(getCnn());
+        
+        String parId = getRequest().getParameter("id");
+        
+        if (!validateInput(parId)) {
+            RequestDispatcher rd = getRequest().getRequestDispatcher("/InputValidationError.jsp");
+            rd.forward(getRequest(), getResponse());
+            return;
+        }
 
         BuildingUnit u = new BuildingUnit();
-        u.setId(Integer.valueOf(getRequest().getParameter("id")));
+        u.setId(Integer.valueOf(parId));
         u.setBuildingId(buildingDao.getBuilding(getCompany().getId()).getId());
         buildingDao.deleteBuildingUnit(u);
         String url = "Dispatcher?page=buildingUnitList";
         getRequest().setAttribute("url", url);
         RequestDispatcher rd = getRequest().getRequestDispatcher("/_refresh.jsp");
         rd.forward(getRequest(), getResponse());
+    }
+    
+    private boolean validateInput(String parId) {
+        boolean result = true;
+        
+        if (!Validator.validateInteger(parId, 0, Validator.maxIntAllowed)) {
+            result = false;
+        }
+        
+        return result;
     }
 }
